pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mai_data"
version = "0.1.0"
description = "MAI Data Utilities - Tools for managing and validating AI model data"
readme = "README.md"
requires-python = ">=3.10"
license = "Apache-2.0"
authors = [
    { name = "Maida AI" }
]
dependencies = [
    "requests>=2.31.0",
    "typer>=0.9.0",
    "hydra-core>=1.3.2",
    "tqdm>=4.66.1",
    "google-cloud-bigquery>=3.17.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-cov>=6.1",
    "ruff>=0.11",
    "black>=25.1",
    "isort>=6.0",
]

# [project.scripts]
# maix = "mai_data.cli:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=mai_data"

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
ignore = [
    "B008",  # Function calls in annotations
]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
isort.known-first-party = ["mai_data"]

[tool.black]
line-length = 120
target-version = ["py310"]

[tool.isort]
profile = "black"
multi_line_output = 3
